Test: unittests for poolcoordinator ip address comparison
